More project restructuring fixes
