+++ backend/src/mirror/plugins/calendars/__init__.py
@@ -1,2 +1,2 @@
-from .calendar import start_plugin
+from .calendars import start_plugin
 from .configure import configure_plugin
